Move AQI conversions out
